Create dan Delete Data

+++ index.php
@@ -77,7 +77,7 @@
                         <td><?php echo $result['alamat']; ?></td>
                         <td>
                             <a href="kelola.php?ubah=<?php echo $result['id_siswa']; ?>" type="button" class="btn btn-success btn-sm"><i class="fa fa-pencil"></i></a>
-                            <a href="proses.php?hapus=<?php echo $result['id_siswa']; ?>" type="button" class="btn btn-danger btn-sm"> <i class="fa fa-trash"></i></a>
+                            <a href="proses.php?hapus=<?php echo $result['id_siswa']; ?>" type="button" class="btn btn-danger btn-sm" onclick="return confirm('Apakah yakin anda ingin menghapus data?')"> <i class="fa fa-trash"></i></a>
                         </td>
                     </tr>
                     <?php
